Test TCN analysis of Lorenti p4679 data -> high R2
A modified version of the driver_load_data.ipynb notebook in the original folder structure provided by Adriano, which implements a TCN after loading sheer data from experiment p4679. TCN train:test split 75:25, inference run on test set and resulting R2 is 98%. Questions: is the TCN implemented correctly? If so, why is R2 so high? Is this a meaningful result or trivial?